Add Caves of Yamar
Caves of Yamar is a text adventure game (in the spirit of Zork) that I wrote in the 1980s using Commodore 64 BASIC.

I extracted this from a C64 into a binary file, and then used VICE to "print" the
the program listing to an ASCII file.

Some of the character set (and case) did not seem to transfer over properly.